start version 0.2

+++ scripts/release-copy.sh
@@ -1,8 +1,16 @@
 #!/usr/bin/env bash
 
 # copy files into GitHub repo so we can push them and make a new release
-LOCAL=$HOME/git/anglican_calendar
+LOCAL=$HOME/git/anglican-calendar
 GITHUB=$HOME/extgit/anglican_calendar
+if [[ ! -d $LOCAL ]]; then
+    echo "cannot find" $LOCAL
+    exit 1
+fi
+if [[ ! -d $GITHUB ]]; then
+    echo "cannot find" $GITHUB
+    exit 1
+fi
 
 echo "deleting targets..."
 rm -rf $LOCAL/*/target
@@ -12,20 +20,15 @@ echo "copying metadata..."
 cp $LOCAL/Cargo.toml $GITHUB
 cp $LOCAL/README.md $GITHUB
 echo "copying src..."
-cp -R $LOCAL/src $GITHUB
+cp -R $LOCAL/calendar $GITHUB
 echo "copying scripts..."
 cp -R $LOCAL/scripts $GITHUB
 echo "copying edit_data..."
 cp -R $LOCAL/edit_data $GITHUB
-#rm -rf $GITHUB/edit_data
-# echo "copying process_data..."
-# cp -R $LOCAL/process_data $GITHUB
 rm -rf $GITHUB/process_data
 echo "copying final data..."
-cp -R $LOCAL/data/final $GITHUB/data
-echo "copying reports..."
-cp -R $LOCAL/data/reports $GITHUB/data
-echo "copying cals ..."
-cp -R $LOCAL/data/cals $GITHUB/data
+cp -R $LOCAL/data/spreads $GITHUB/data
+# echo "copying cals ..."
+# cp -R $LOCAL/data/cals $GITHUB/data
 
 echo "make sure that files are tagged"
